% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zohor_get_report.R
\name{zohor_get_report}
\alias{zohor_get_report}
\title{Downloads report from Zoho creator}
\usage{
zohor_get_report(zohor_report, ...)
}
\arguments{
\item{zohor_report}{A string. name of the report (link name) in Zoho creator}
}
\description{
Downloads report from Zoho creator
}
