% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zohor_login.R
\name{zohor_login}
\alias{zohor_login}
\title{Define API credentials for your Zoho creator app}
\usage{
zohor_login(
  zohor_server,
  zohor_user,
  zohor_ownername,
  zohor_scope,
  zohor_token,
  zohor_password,
  ...
)
}
\arguments{
\item{zohor_server}{A string. URL of your App.}

\item{zohor_user}{A string. email address of
your user in Zoho creator.}

\item{zohor_ownername}{A string. Owner of the app}

\item{zohor_password}{A string. Password of the API user.}

\item{zoho_scope}{A string. Scope of the query. Default = "creatorapi"}

\item{zoho_token}{A string. Token of the API}
}
\value{
All the parameters provided to the function (\code{zohor_server},
\code{zohor_user},  \code{zohor_password}), etc. will be saved in your global environment so can be used
in your workflow. A
}
\description{
#'\code{zohor_login()} saves these credentials in the global environment so
they can be used to communicate with the server
This function must be run before any other function of this
package \code{zohor}.
}
\examples{
zohor_login(zohor_server = "https://creator.zoho.com/api/json/myapp/view/",
              zoho_user = "ronaldo",
               zohor_ownername = "ronaldogmail.com,
               zoho_scope = "creatorapi,
               zoho_token = XXXXXXXXX,
               zohor_password = golazosolitario,
             )



}
